# src/main/resources/api-docs.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8050
    description: Generated server url
paths:
  /api/uploads:
    post:
      tags:
        - uploads-controller
      operationId: fileUpload
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                picture:
                  $ref: '#/components/schemas/Picture'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Picture'
  /api/auth/signup:
    post:
      tags:
        - auth-controller
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/auth/signin:
    post:
      tags:
        - auth-controller
      operationId: authenticateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/admin/uploads/{pictureId}:
    get:
      tags:
        - admin-uploads-controller
      operationId: getUploadById
      parameters:
        - name: pictureId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdminUploadDTO'
    patch:
      tags:
        - admin-uploads-controller
      operationId: updateStatus
      parameters:
        - name: pictureId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
  /api/uploads/{pictureId}:
    get:
      tags:
        - uploads-controller
      operationId: getUploadsById
      parameters:
        - name: pictureId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserUploadDTO'
  /api/uploads/accepted:
    get:
      tags:
        - uploads-controller
      operationId: getAcceptedPictures
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
  /api/admin/uploads/unprocessed:
    get:
      tags:
        - admin-uploads-controller
      operationId: getAcceptedPictures_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: array
                  items:
                    type: string
                    format: byte
components:
  schemas:
    Picture:
      type: object
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
        description:
          type: string
        attachment:
          type: array
          items:
            type: string
            format: byte
        status:
          type: string
          enum:
            - ACCEPTED
            - REJECTED
            - UNPROCESSED
        urls:
          type: string
        category:
          type: string
          enum:
            - livingThing
            - machine
            - nature
        user:
          $ref: '#/components/schemas/User'
        originalFilename:
          type: string
    Role:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          enum:
            - ROLE_USER
            - ROLE_ADMIN
    User:
      required:
        - email
        - password
        - username
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          maxLength: 20
          minLength: 0
          type: string
        email:
          maxLength: 50
          minLength: 0
          type: string
        password:
          maxLength: 120
          minLength: 0
          type: string
        roles:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Role'
        picture:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Picture'
    SignupRequest:
      required:
        - email
        - password
        - username
      type: object
      properties:
        username:
          maxLength: 20
          minLength: 3
          type: string
        email:
          maxLength: 50
          minLength: 0
          type: string
        role:
          uniqueItems: true
          type: array
          items:
            type: string
        password:
          maxLength: 40
          minLength: 6
          type: string
    LoginRequest:
      required:
        - password
        - username
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    UserUploadDTO:
      type: object
      properties:
        category:
          type: string
          enum:
            - livingThing
            - machine
            - nature
        attachment:
          type: array
          items:
            type: string
            format: byte
        description:
          type: string
    AdminUploadDTO:
      type: object
      properties:
        width:
          type: integer
          format: int32
        height:
          type: integer
          format: int32
        category:
          type: string
          enum:
            - livingThing
            - machine
            - nature
        attachment:
          type: array
          items:
            type: string
            format: byte
        description:
          type: string